Module sttp::transport

Represents transport functionality of the STTP library.

Structs

CompressionModesDeprecated

Enumeration of the possible compression modes supported by STTP.

DataPacketFlags

Defines the bitflags representing the possible flags for a data packet.

Defaults

Defines default values for various STTP settings.

Measurement

Represents a basic unit of measured data for transmission or reception in the STTP API.

OperationalModes

Enumeration of the possible modes that affect how DataPublisher and DataSubscriber communicate during as STTP session.

SignalIndexCache

Represents a mapping of 32-bit runtime IDs to 128-bit globally unique measurement IDs. The class additionally provides reverse lookup and an extra mapping to human-readable measurement keys.

StateFlags

Defines the bitflags representing the possible quality states of a Measurement value.

Enums

ConnectStatus

Enumeration of the possible connection status results used by the SubscriberConnector.

OperationalEncoding

Enumeration of the possible string encoding options of an STTP session.

SecurityMode

Enumeration of the possible security modes used by the DataPublisher to secure data sent over the command channel in STTP.

ServerCommand

Enumeration of the possible server commands received by a DataPublisher and sent by a DataSubscriber during an STTP session.

ServerResponse

Enumeration of the possible server responses sent by DataPublisher and received by DataSubscriber during an STTP session.