Struct sttp::transport::CompactMeasurement

pub struct CompactMeasurement { /* private fields */ }

Represents a compact Measurement for transmission or reception in the STTP API.

Implementations

impl CompactMeasurement

pub fn new(
signal_index_cache: Arc<SignalIndexCache>,
include_time: bool,
use_millisecond_resolution: bool
) -> Self

Creates a new CompactMeasurement.

pub fn include_time(&self) -> bool

Gets flag that determines if time is serialized into measurement binary image.

pub fn get_binary_length(&mut self) -> usize

Gets the length of the CompactMeasurement.

pub fn get_timestamp_c2(&self) -> u16

Gets offset compressed millisecond-resolution 2-byte timestamp.

pub fn get_timestamp_c4(&self) -> u32

Gets offset compressed tick-resolution 4-byte timestamp.

pub fn get_compact_state_flags(&self) -> u8

Gets byte level compact state flags with encoded time index and base time offset bits.

pub fn set_compact_state_flags(&mut self, value: u8)

Sets byte level compact state flags with encoded time index and base time offset bits.

pub fn get_runtime_id(&self) -> i32

Gets the 4-byte run-time signal index for this measurement.

pub fn set_runtime_id(&mut self, signal_index: i32)

Sets the 4-byte run-time signal index for this measurement.

pub fn decode(&mut self, buffer: &[u8]) -> Result<usize, Box<dyn Error>>

parses a CompactMeasurement from the specified byte buffer.