Struct sttp::transport::Defaults

pub struct Defaults;

Defines default values for various STTP settings.

Implementations

impl Defaults

pub const MAX_RETRIES: i32 = -1i32

Default for maximum number of retries for a connection attempt.

pub const RETRY_INTERVAL: f64 = 1f64

Default for retry interval in seconds.

pub const MAX_RETRY_INTERVAL: f64 = 30f64

Default for maximum retry interval in seconds.

pub const AUTO_RECONNECT: bool = true

Default for auto-reconnect flag.

pub const AUTO_REQUEST_METADATA: bool = true

Default for auto-request metadata flag.

pub const AUTO_SUBSCRIBE: bool = true

Default for auto-subscribe flag.

pub const COMPRESS_PAYLOAD_DATA: bool = true

Default for compress payload data flag.

pub const COMPRESS_METADATA: bool = true

Default for compress metadata flag.

pub const COMPRESS_SIGNAL_INDEX_CACHE: bool = true

Default for compress signal index cache flag.

pub const METADATA_FILTERS: &str = ""

Default for metadata filters.

pub const SOCKET_TIMEOUT: f64 = 2f64

Default for socket timeout in seconds.

pub const VERSION: u8 = 2u8

Default for STTP version.

pub const FILTER_EXPRESSION: &str = ""

Default for filter expression.

pub const THROTTLED: bool = false

Default for throttled flag.

pub const PUBLISH_INTERVAL: f64 = 1f64

Default for publish interval in seconds.

pub const ENABLE_UDP_DATA_CHANNEL: bool = false

Default for enable UDP data channel flag.

pub const DATA_CHANNEL_LOCAL_PORT: u16 = 0u16

Default for local port for data channel.

pub const DATA_CHANNEL_INTERFACE: &str = ""

Default for interface for data channel.

pub const INCLUDE_TIME: bool = true

Default for include time flag.

pub const ENABLE_TIME_REASONABILITY_CHECK: bool = false

Default for enable time reasonability check flag.

pub const LAG_TIME: f64 = 10f64

Default for lag time in seconds.

pub const LEAD_TIME: f64 = 5f64

Default for lead time in seconds.

pub const USE_LOCAL_CLOCK_AS_REAL_TIME: bool = false

Default for use local clock as real time flag.

pub const USE_MILLISECOND_RESOLUTION: bool = false

Default for use millisecond resolution flag.

pub const REQUEST_NAN_VALUE_FILTER: bool = false

Default for request NAN-value filter flag.

pub const START_TIME: &str = ""

Default for start time.

pub const STOP_TIME: &str = ""

Default for stop time.

pub const CONSTRAINT_PARAMETERS: &str = ""

Default for constraint parameters.

pub const PROCESSING_INTERVAL: i32 = -1i32

Default for processing interval in seconds.

pub const EXTRA_CONNECTION_STRING_PARAMETERS: &str = ""

Default for extra connection string parameters.