Struct sttp::transport::SubscriptionInfo

pub struct SubscriptionInfo {

    pub filter_expression: String,
    pub throttled: bool,
    pub publish_interval: f64,
    pub enable_udp_data_channel: bool,
    pub data_channel_local_port: u16,
    pub data_channel_interface: String,
    pub include_time: bool,
    pub use_millisecond_resolution: bool,
    pub request_nan_value_filter: bool,
    pub start_time: String,
    pub stop_time: String,
    pub constraint_parameters: String,
    pub processing_interval: i32,
    pub extra_connection_string_parameters: String,

}

Defines subscription related settings for a DataSubscriber instance.

Fields
filter_expression: String

Gets or sets the desired measurements for a subscription. Examples include:

Directly specified signal IDs (UUID values in string format):

38A47B0-F10B-4143-9A0A-0DBC4FFEF1E8; {E4BBFE6A-35BD-4E5B-92C9-11FF913E7877}

Directly specified tag names:

DOM_GPLAINS-BUS1:VH; TVA_SHELBY-BUS1:VH

A filter expression against a selection view:

FILTER ActiveMeasurements WHERE Company = 'GPA' AND SignalType = 'FREQ'

throttled: bool

Gets or sets flag that determines if data will be published using down-sampling.

publish_interval: f64

Gets or sets the down-sampling publish interval to use when Throttled is true.

enable_udp_data_channel: bool

Gets or sets flag that requests that a UDP channel be used for data publication.

data_channel_local_port: u16

Gets or sets the desired UDP port to use for publication.

data_channel_interface: String

Gets or sets the desired network interface to use for UDP publication.

include_time: bool

Gets or sets flag that determines if time should be included in non-compressed, compact measurements.

use_millisecond_resolution: bool

Gets or sets flag that determines if time should be restricted to milliseconds in non-compressed, compact measurements.

request_nan_value_filter: bool

Gets or sets flag that requests that the publisher filter, i.e., does not send, any NaN values.

start_time: String

Gets or sets that defines the start time for a requested temporal data playback, i.e., a historical subscription. Simply by specifying a StartTime and StopTime, a subscription is considered a historical subscription. Note that the publisher may not support historical subscriptions, in which case the subscribe will fail.

stop_time: String

Gets or sets that defines the stop time for a requested temporal data playback, i.e., a historical subscription. Simply by specifying a StartTime and StopTime, a subscription is considered a historical subscription. Note that the publisher may not support historical subscriptions, in which case the subscribe will fail.

constraint_parameters: String

Gets or sets any custom constraint parameters for a requested temporal data playback. This can include parameters that may be needed to initiate, filter, or control historical data access.

processing_interval: i32

Gets or sets the initial playback speed, in milliseconds, for a requested temporal data playback. With the exception of the values of -1 and 0, this value specifies the desired processing interval for data, i.e., basically a delay, or timer interval, over which to process data.A value of -1 means to use the default processing interval while a value of 0 means to process data as fast as possible.

extra_connection_string_parameters: String

Gets or sets any extra or custom connection string parameters that may be needed for a subscription.